test(operator): isolate ownership conflict conditions
